+++ src/operations/func.py
@@ -49,9 +49,9 @@ def run(self, op: dict):
     # declare function
     is_method = False
     try:
-        self.current_struct
+        self.current_class
         self.current_func = arg
-        self.structs[self.current_struct].methods[self.current_func] = []
+        self.classes[self.current_class].methods[self.current_func] = []
         is_method = True
     except:
         self.current_func = self.current_namespace() + arg
@@ -62,6 +62,6 @@ def run(self, op: dict):
         arg_var = op['args'][1].strip(')').strip('(')
         self.arg_should_be_variable(arg_var, op)
         if is_method:
-            self.structs[self.current_struct].methods[self.current_func].append(parser.parse(arg_var + ' = ^', '<system>')[0])
+            self.classes[self.current_class].methods[self.current_func].append(parser.parse(arg_var + ' = ^', '<system>')[0])
         else:
             self.functions[self.current_func].append(parser.parse(arg_var + ' = ^', '<system>')[0])